requirements.txt
streamlit~=1.40.1
pandas~=2.0.3
requests~=2.32.3
shap~=0.44.1
numpy
groq~=0.20.0
bcrypt~=4.3.0
python-dotenv~=1.0.1
lightgbm
rpds-py
pyarrow
scipy~=1.10.1